+++ src/components/Gantt.vue
@@ -3,9 +3,13 @@ import { computed, ref } from 'vue'
 import type {
   GanttCellEvent,
   GanttColumnEvent,
+  GanttDependencyChange,
   GanttDependencyEvent,
+  GanttDependencyUpdate,
   GanttGroupToggleEvent,
   GanttMoveEvent,
+  GanttProgressEvent,
+  GanttResizeEvent,
   GanttRootProps,
   GanttRowEvent,
   GanttScrollOptions,
@@ -21,7 +25,12 @@ const props = defineProps<GanttRootProps & {
 
 const emit = defineEmits<{
   move: [event: GanttMoveEvent]
+  resize: [event: GanttResizeEvent]
+  progress: [event: GanttProgressEvent]
   'group-toggle': [event: GanttGroupToggleEvent]
+  'dependency-create': [event: GanttDependencyChange]
+  'dependency-remove': [event: GanttDependencyChange]
+  'dependency-update': [event: GanttDependencyUpdate]
   'task-click': [event: GanttTaskEvent]
   'task-dblclick': [event: GanttTaskEvent]
   'task-contextmenu': [event: GanttTaskEvent]
@@ -75,7 +84,12 @@ defineExpose({
     ref="root"
     v-bind="rootProps"
     @move="emit('move', $event)"
+    @resize="emit('resize', $event)"
+    @progress="emit('progress', $event)"
     @group-toggle="emit('group-toggle', $event)"
+    @dependency-create="emit('dependency-create', $event)"
+    @dependency-remove="emit('dependency-remove', $event)"
+    @dependency-update="emit('dependency-update', $event)"
     @task-click="emit('task-click', $event)"
     @task-dblclick="emit('task-dblclick', $event)"
     @task-contextmenu="emit('task-contextmenu', $event)"